Revert back to .NET Framework
And use v4.6.2 to maximize portability as it's the earliest supported version (https://docs.microsoft.com/en-US/lifecycle/faq/dotnet-framework).

The only features we were missing are the init setters in Monitor.cs, but that's an easy change.

File size is under 400KB. On self-contained .NET Core it's over 150MB! Without trimming (https://github.com/dotnet/wpf/issues/3811) on self-contained deployment, it's very hard to justify the file size, and not being self-contained isn't an option as Windows doesn't come with a system install of .NET Core.